requirements.txt
tensorflow
numpy
opencv-python
gdown
tqdm
PyYAML
dvc[gdrive]
Flask